requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.24.0

# Machine learning libraries
scikit-learn>=1.3.0
xgboost>=1.7.0

# Financial data
yfinance>=0.2.0

# Web requests and API handling
requests>=2.28.0

# Natural language processing
textblob>=0.17.0

# Environment variable management
python-dotenv>=1.0.0

# Web app framework
Streamlit>=1.28.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
ploty>=5.15.0
